# swagger.yaml
openapi: 3.0.0
info:
  title: Q&A Platform API
  version: 1.0.0
  description: REST API for a StackOverflow-like Q&A platform
servers:
  - url: http://localhost:3000
paths:
...
  /answers/questions/{id}/answers:
    post:
      summary: Answer a question
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerCreate'
      responses:
        '201':
          description: Answer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerWithVoteStatus'
  /answers/{id}/mark-correct:
    patch:
      summary: Mark answer as correct
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Answer marked as correct
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerWithVoteStatus'
  /answers/{id}/statistics:
    get:
      summary: Get answer stats (total votes)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Answer statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalVotes:
                    type: integer
  /answers/{id}/vote:
    post:
      summary: Vote on an answer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: integer
                  enum: [1, -1]
      responses:
        '200':
          description: Vote registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerWithVoteStatus'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnswerCreate:
      type: object
      properties:
        content:
          type: string
    Answer:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        questionId:
          type: integer
        userId:
          type: integer
        isCorrect:
          type: boolean
        correct:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AnswerWithVoteStatus:
      allOf:
        - $ref: '#/components/schemas/Answer'
        - type: object
          properties:
            hasVoted:
              type: boolean 